Refactor handle types.
Change all the handles to tuple structs containing a void
pointer. These have the same representation as a bare pointer, but are
separate types so the compiler will stop us from using the wrong
handle type.

The handles intentionally don't have any direct connection to the type
they point at. In the long run the types submodule may be spun out
into a more generic OFX support crate so it shouldn't be tied to our
specific implementations of host objects.

In the main crate, use traits to connect handles to the objects and
make them convenient to work with.