Add definitions of OFX string constants.
These were done by copy-paste plus some text munging, and there are a
few irregularities in the original definitions (e.g. string values not
matching identifiers), so there may be errors in properties which
haven't been tested yet.

I made a first pass at grouping them into submodules. Not sure if this
is the right organization, or worth it in the long run.