Refactor handles again.
Keep the handle types the same, but rewrite the traits in the main
create to work with Arc<Mutex<T>> objects and avoid dereferencing raw
pointers.

This version is thread safe, works with objects stored in
containers (which could get moved around if not boxed), and can detect
invalid or stale handles.