Switch OpenEXR libraries.
The openexr crate is a wrapper around the official reference
implementation. It has to load the libary at runtime, which sometimes
required mucking with LD_LIBRARY_PATH.

Instead, use the exr crate which is pure Rust with no unsafe code. It
doesn't require any external libraries and the API is a bit more
streamlined for our purposes.