Store library in LoadedPlugin to keep it loaded.
When libloading::Library is dropped, it unloads the library even if we
still have references to functions in it. For some reason this didn't
cause problems until now -- it looks like dlclose() queues things to
be unloaded by a background thread, so maybe there was enough to delay
to get away with it in most cases.