Improve handling of string constants.
Introduce an OfxStr type which wraps a byte slice that is compatible
with both Rust and C strings. Invariants are checked at construction
time, so later conversions are safe and infallible.

Use OfxStr for all constants, and everywhere that strings cross the
API boundary.

This eliminates a lot of verbose string conversions.