Add full layout options to RenderFilter command.
Includes project dims and input image origin as well as render
window. Grouping them in a struct ensures that all three are provided
together. This sidesteps the thorny problem of choosing sensible
defaults when only one or two are provided.

If the layout is missing, everything is based off the size
of the input image.